# src/docs/swagger.yaml
basePath: /api/
definitions:
  dto.GetUserRequest:
    properties:
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - mobile_number
    type: object
...
  helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/validations.ValidationError'
        type: array
    type: object
  helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
host: localhost:5005
info:
  contact:
    email: rezam578@gmail.com
    name: Reza Mobaraki
    url: https://linkedin.com/in/reza-mobaraki
  description: This is a sample Golang server CarApp server with gin
  title: CarApp API
  version: "1.0"
paths:
...
  /v1/users/send-otp/:
    post:
      consumes:
      - application/json
      description: Send OTP to user
      parameters:
      - description: User details for OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: Send OTP
      tags:
      - User
schemes:
- http
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"